Vulkan: Avoid storage reallocation in UpdateDescriptorSetsBuilder
UpdateDescriptorSetsBuilder::mDescriptorBufferInfos and
mDescriptorImageInfos will keep grow to a few hundreds of entries and
that grow will end up with data copy and patching mWriteDescriptorSets.
There is no requirement that entire vector of mDescriptorBufferInfos
andmDescriptorImageInfos must be continuous. The only requirement is
that when allocDescriptorBufferInfos(count) is called, the count of
entries must be continuous. This CL uses a queue of vectors so that when
we need to allocate new storage we just add another vector and allocate
out of the new vector. This avoids all related data copy.

Similar thing applies to mWriteDescriptorSets. The only thing I added
for mWriteDescriptorSets is that I try to grow the first vector big
enough to hold all of the entries for next submission to minimize the
vkUpdateDescriptorSets call.

Bug: b/293297177
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6253675
Commit-Queue: Charlie Lao <cclao@google.com>

+++ src/libANGLE/renderer/vulkan/ProgramExecutableVk.cpp
@@ -265,7 +265,7 @@ angle::Result UpdateFullTexturesDescriptorSet(vk::ErrorContext *context,
 
     // Allocate VkWriteDescriptorSet and initialize the data structure
     VkWriteDescriptorSet *writeDescriptorSets =
-        updateBuilder->allocWriteDescriptorSets(writeDescriptorDescs.size());
+        updateBuilder->allocWriteDescriptorSets(static_cast<uint32_t>(writeDescriptorDescs.size()));
     for (uint32_t writeIndex = 0; writeIndex < writeDescriptorDescs.size(); ++writeIndex)
     {
         ASSERT(writeDescriptorDescs[writeIndex].descriptorCount > 0);
